fix: remove plugin approach, first shot at fixing composable
