+++ playground/pages/index.vue
@@ -6,10 +6,10 @@
     <div ref="pdfSection">
       Thanks for testing out nuxt-pdf!
     </div>
-    <button @click="print(pdfSection, {})">
+    <button @click="print(pdfSection)">
       print card
     </button>
-    <button @click="print(page, {})">
+    <button @click="print(page)">
       print page
     </button>
   </div>
@@ -17,13 +17,15 @@
 
 <script setup lang="ts">
 import { ref } from 'vue'
-import { Html2PdfOptions } from '../../src/runtime/plugin.client'
 import { usePDFExport } from '#imports'
 
 const page = ref<HTMLElement | null>(null)
 const pdfSection = ref<HTMLElement | null>(null)
 
-const print = async (element: HTMLElement, options: Html2PdfOptions) => {
+const print = async (element: HTMLElement | null, options?: Parameters<typeof usePDFExport>[1]) => {
+  if (!element) {
+    return
+  }
   await usePDFExport(element, options)
 }
 </script>
